03.Django/03.reproduce/toyproject/requirements.txt
Django
djangorestframework
psycopg2
gunicorn
supervisor
Pillow
redis
rq
django-rq
drf-yasg
